fix: make restOfArgs work with raw params, ct check, fixes
revert back the getopt args/behavior to not accept
`--param value` (with space instead of e.g. `=`:
`--param=value`) as this was breaking things for
boolean flags: in `--flag1 --flag2` the lib
expects `--flag2` is the value for `--flag1`
the space behavior is useful for codetracer, but the
flag one is more important, maybe we can special case
flags?

it seemed more correct to work with direct params, without
any further processing and to make sure
there can be only required arguments before restOfArgs
and that it is last, so we can more easily apply
this param logic
as otherwise there is no builtin order like
<flags> <argument> so i am not sure how would we detect
when it's time for restOfArgs

this was needed to fix/workaround problems where
codetracer wrapping commands pass many params
to other tools/compilers/linkers